requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.10.2
gunicorn==19.9.0
icalendar==4.0.3
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.3
